talentmap/requirements.txt
flask
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy==2.*
email_validator
python-decouple
gunicorn
requests
json2html
WTForms==2.3.3
SQLAlchemy==1.3.20
plotly==5.5.0 
pandas==1.2.5
shap==0.39.0 
numpy==1.21
numba
